+++ new-dti-website-redesign/src/app/course/PastStudentProjects.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import Image from 'next/image';
-import FancyTabs from '../FancyTabs/FancyTabs';
+import FancyTabs from '../../components/FancyTabs/FancyTabs';
 import studentProjectData from '../../app/course/data/student_projects.json';
-import Button from '../Button';
+import Button from '../../components/Button';
 
 type Project = {
   title: string;
@@ -12,7 +12,7 @@ type Project = {
   link?: string;
 };
 
-export default function StudentProjectsSection() {
+export default function PastStudentProjects() {
   const { student_projects }: { student_projects: Project[] } = studentProjectData;
 
   const tabs = student_projects.map((project) => ({
@@ -38,5 +38,16 @@ export default function StudentProjectsSection() {
     )
   }));
 
-  return <FancyTabs tabs={tabs} />;
+  return (
+    <section>
+      <div className="flex flex-col gap">
+        <div className="p-4 sm:p-8 pb-8 md:pb-0">
+          <h2>Past student projects</h2>
+        </div>
+        <div className="p-0 md:p-8">
+          <FancyTabs tabs={tabs} />
+        </div>
+      </div>
+    </section>
+  );
 }
