+++ new-dti-website-redesign/src/app/products/ProductsList.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import products from './products.json';
+import Product from './Product';
+import SectionSep from '../../components/SectionSep';
+
+export default function ProductsList() {
+  return (
+    <>
+      {products.map((product, index) => (
+        <React.Fragment key={product.name}>
+          <Product {...product} />
+          {index < products.length - 1 && (
+            <SectionSep grid className="border-x-1 border-border-1" />
+          )}
+        </React.Fragment>
+      ))}
+    </>
+  );
+}
