+++ new-dti-website-redesign/src/app/sponsor/SponsorFeatures.tsx
@@ -1,0 +1,73 @@
+import React from 'react';
+import FeatureCard from '../../components/FeatureCard';
+
+export default function SponsorFeatures() {
+  return (
+    <section className="grid grid-cols-1 md:grid-cols-3">
+      <FeatureCard
+        title="Build relationships"
+        body="We help organizations create a diverse talent pipeline, present information sessions, conduct workshops, and help establish a presence on Cornell's campus."
+        icon={
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+            stroke="currentColor"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            className="lucide lucide-heart-icon lucide-heart"
+          >
+            <path d="M19 14c1.49-1.46 3-3.21 3-5.5A5.5 5.5 0 0 0 16.5 3c-1.76 0-3 .5-4.5 2-1.5-1.5-2.74-2-4.5-2A5.5 5.5 0 0 0 2 8.5c0 2.3 1.5 4.05 3 5.5l7 7Z" />
+          </svg>
+        }
+      />
+      <FeatureCard
+        title="Help us thrive"
+        body="Your contributions are crucial in helping the DTI community grow and maintain its vibrant culture. You can make a direct impact on the Cornell experience!"
+        icon={
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+            stroke="currentColor"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            className="lucide lucide-trending-up-icon lucide-trending-up"
+          >
+            <path d="M16 7h6v6" />
+            <path d="m22 7-8.5 8.5-5-5L2 17" />
+          </svg>
+        }
+      />
+      <FeatureCard
+        title="Make an impact"
+        body="Through software licenses, community outreach, marketing, and more, funds help us grow our vision of helping our the community."
+        icon={
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            fill="none"
+            stroke="currentColor"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            className="lucide lucide-earth-icon lucide-earth"
+          >
+            <path d="M21.54 15H17a2 2 0 0 0-2 2v4.54" />
+            <path d="M7 3.34V5a3 3 0 0 0 3 3a2 2 0 0 1 2 2c0 1.1.9 2 2 2a2 2 0 0 0 2-2c0-1.1.9-2 2-2h3.17" />
+            <path d="M11 21.95V18a2 2 0 0 0-2-2a2 2 0 0 1-2-2v-1a2 2 0 0 0-2-2H2.05" />
+            <circle cx="12" cy="12" r="10" />
+          </svg>
+        }
+      />
+    </section>
+  );
+}
